Create change
GitHub Pull Request: https://github.com/dstanek/snake-guice/pull/2

Fixes for google issues 11, 14, and 16

Apologies for the messy commit history in the pull request.



I would have rebased but you're not supposed to do that for published history (i.e. google code).



You may notice I reverted the provider instance caching.

I don't think you can do that without verifying that the provider exists in the current binder's (or parent's) Singleton scope instance. (The fix for issue 11.)



The fix for issue 11 essentially means you need to register a scope to the binder by calling binder.bind_scope before you can bind something into the scope.



The binder.get_binding change if I recall correctly was to simplify life in the multibinder so that the multibinder class didn't have constructed Key instances. (Since the keys are so complex in multibinder.)



The reason for the complexity is that all items in the multibinding have individual bindings in the binder (as well as a dynamic provider to inject the multibinder).



The key of the dynamic provider class is:



    Key(<_MultiBinderSubClass>.multibinding_type(<userSuppliedClass>),

           <userSuppliedAnnotation>)



The key of each of the multibinder items is:



    Key(<userSuppliedClass>, 

           <unique _RealElement(

                           str(<userSuppliedAnnotation>), # the name 

                           <_MultiBinderSubClass>.multibinding_type.__name__) # the type

             instance>)



The core concept of these keys is _RealElement which is essentially:



     (<unique integer across all instances of _RealElement>, 

       <name>, <type>)



The new _ _ str _ _  method for Annotation is just as important since it becomes the name of _RealElement.



Thanks again for snake-guice,

Bill



Patch-set: 1
Change-id: I735c209e67db889975fed251e3bd24cb20710262
Subject: Fix issue16 in the main snake guice repo.
Branch: refs/heads/master
Commit: 5abad320fe7156162c71cc0772b4674bb490e840
Groups: ecc9d14a639d245642e8fd05535a0c93d388419e